added the updated version of calculator

+++ Calculator/src/Calculator1/Calculator2.java
@@ -3,23 +3,23 @@ package Calculator1;
 public class Calculator2 {
 	
 
-	public static int add (int num1, int num2) {
-		int sum = num1 + num2;
+	public static double add (double num1, double num2) {
+		double sum = num1 + num2;
 		return sum; 
 	}
 	
-	public static int subtract (int num1, int num2) {
-		int sum = num1 - num2;
+	public static double subtract (double num1, double num2) {
+		double sum = num1 - num2;
 		return sum; 
 	}
 	
-	public static int multiply (int num1, int num2) {
-		int sum = num1 * num2;
+	public static double multiply (double num1, double num2) {
+		double sum = num1 * num2;
 		return sum; 
 	}
 	
-	public static float divide (float num1, float num2) {
-		float sum = num1 / num2;
+	public static double divide (double num1, double num2) {
+		double sum = num1 / num2;
 		if(num1 > num2) {
             System.out.println(sum);
         } else {
